Align ImportRun domain layer with #41907 refinement
Following the refinement on #41907:
- Remove UpdateImportRunColumnMapping (folded into StartImportRun, which now
  also carries the batch-size limit)
- Drop the four ImportMatch CQRS messages + their handlers/DTO/exceptions
  (saved-mapping presets stay BO-only on the existing ImportMatchRepository)
- Merge GetImportRunForEditing + GetImportRunProgress into a single
  GetImportRunState query returning an ImportRunState DTO
- RunImportBatch payload reduced to runId (offset/limit now read from the
  persisted run)
