--- src/Core/Domain/Import/Exception/ImportMatchNotFoundException.php
@@ -1,16 +1,0 @@
-<?php
-/**
- * For the full copyright and license information, please view the
- * docs/licenses/LICENSE.txt file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace PrestaShop\PrestaShop\Core\Domain\Import\Exception;
-
-/**
- * Thrown when a saved data-matching configuration cannot be found.
- */
-final class ImportMatchNotFoundException extends ImportException
-{
-}
