--- src/Core/Domain/Import/Query/GetImportMatches.php
@@ -1,16 +1,0 @@
-<?php
-/**
- * For the full copyright and license information, please view the
- * docs/licenses/LICENSE.txt file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace PrestaShop\PrestaShop\Core\Domain\Import\Query;
-
-/**
- * Gets all saved data-matching configurations (to populate the load dropdown on step 2).
- */
-final class GetImportMatches
-{
-}
