--- src/Core/Domain/Import/QueryHandler/GetImportMatchesHandlerInterface.php
@@ -1,23 +1,0 @@
-<?php
-/**
- * For the full copyright and license information, please view the
- * docs/licenses/LICENSE.txt file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace PrestaShop\PrestaShop\Core\Domain\Import\QueryHandler;
-
-use PrestaShop\PrestaShop\Core\Domain\Import\Query\GetImportMatches;
-use PrestaShop\PrestaShop\Core\Domain\Import\QueryResult\ImportMatch;
-
-/**
- * Defines the contract for handling @see GetImportMatches.
- */
-interface GetImportMatchesHandlerInterface
-{
-    /**
-     * @return ImportMatch[]
-     */
-    public function handle(GetImportMatches $query): array;
-}
