--- src/Core/Domain/Import/QueryResult/ImportRunProgress.php
@@ -1,129 +1,0 @@
-<?php
-/**
- * For the full copyright and license information, please view the
- * docs/licenses/LICENSE.txt file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace PrestaShop\PrestaShop\Core\Domain\Import\QueryResult;
-
-/**
- * Progress snapshot of an import run, used to poll the progress modal.
- */
-final class ImportRunProgress
-{
-    /**
-     * @var string
-     */
-    private $importRunId;
-
-    /**
-     * @var string
-     */
-    private $status;
-
-    /**
-     * @var int
-     */
-    private $doneCount;
-
-    /**
-     * @var int
-     */
-    private $totalCount;
-
-    /**
-     * @var bool
-     */
-    private $finished;
-
-    /**
-     * @var string[]
-     */
-    private $errors;
-
-    /**
-     * @var string[]
-     */
-    private $warnings;
-
-    /**
-     * @var string[]
-     */
-    private $notices;
-
-    /**
-     * @param string[] $errors
-     * @param string[] $warnings
-     * @param string[] $notices
-     */
-    public function __construct(
-        string $importRunId,
-        string $status,
-        int $doneCount,
-        int $totalCount,
-        bool $finished,
-        array $errors = [],
-        array $warnings = [],
-        array $notices = []
-    ) {
-        $this->importRunId = $importRunId;
-        $this->status = $status;
-        $this->doneCount = $doneCount;
-        $this->totalCount = $totalCount;
-        $this->finished = $finished;
-        $this->errors = array_values($errors);
-        $this->warnings = array_values($warnings);
-        $this->notices = array_values($notices);
-    }
-
-    public function getImportRunId(): string
-    {
-        return $this->importRunId;
-    }
-
-    public function getStatus(): string
-    {
-        return $this->status;
-    }
-
-    public function getDoneCount(): int
-    {
-        return $this->doneCount;
-    }
-
-    public function getTotalCount(): int
-    {
-        return $this->totalCount;
-    }
-
-    public function isFinished(): bool
-    {
-        return $this->finished;
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getErrors(): array
-    {
-        return $this->errors;
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getWarnings(): array
-    {
-        return $this->warnings;
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getNotices(): array
-    {
-        return $this->notices;
-    }
-}
